Broke up the monolithic matchers file into its sub-parts.
