+++ kakao/src/main/kotlin/com/agoda/kakao/matchers/TextInputLayoutCounterEnabledMatcher.kt
@@ -1,0 +1,21 @@
+package com.agoda.kakao.matchers
+
+import android.support.design.widget.TextInputLayout
+import android.view.View
+import org.hamcrest.Description
+import org.hamcrest.TypeSafeMatcher
+
+class TextInputLayoutCounterEnabledMatcher(private val enabled: Boolean) : TypeSafeMatcher<View>() {
+    override fun matchesSafely(item: View?): Boolean {
+        return item?.let {
+            if (item is TextInputLayout) {
+                item.isCounterEnabled == enabled
+            } else false
+        } ?: false
+    }
+
+    override fun describeTo(desc: Description) {
+        desc.appendText("with counter state: ")
+                .appendValue(enabled)
+    }
+}
